# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/products": {
      "get": {
        "description": "Par defaut, la liste est allegee (sans variantes). Ajoutez includeVariants=true pour inclure les variantes. Ajoutez fields=compact pour un payload minimum (id, ref, nom, statut, prix).",
        "operationId": "ProductsController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Recherche textuelle",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filtre de statut",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "Filtre de type / origine / role selon le module",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dateFrom",
            "required": false,
            "in": "query",
            "description": "Date debut (ISO). Champ cible selon le module (createdAt, orderDate, issueDate...).",
            "schema": {
              "example": "2026-01-01T00:00:00.000Z",
              "type": "string"
            }
          },
          {
            "name": "dateTo",
            "required": false,
            "in": "query",
            "description": "Date fin (ISO). Champ cible selon le module (createdAt, orderDate, issueDate...).",
            "schema": {
              "example": "2026-12-31T23:59:59.999Z",
              "type": "string"
            }
          },
          {
            "name": "clientId",
            "required": false,
            "in": "query",
            "description": "Filtre client (avance)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ownership",
            "required": false,
            "in": "query",
            "description": "Filtre propriété produit (COMPANY = catalogue, CLIENT = modèle client). Utilisé avec clientId = ownerClientId pour les modèles d'un client.",
            "schema": {
              "type": "string",
              "enum": [
                "COMPANY",
                "CLIENT"
              ]
            }
          },
          {
            "name": "supplierId",
            "required": false,
            "in": "query",
            "description": "Filtre fournisseur (avance)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "description": "Filtre categorie produit (avance)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "productId",
            "required": false,
            "in": "query",
            "description": "Filtre produit (avance)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "salesOrderId",
            "required": false,
            "in": "query",
            "description": "Filtre commande de vente (avance)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "componentId",
            "required": false,
            "in": "query",
            "description": "Filtre composant / matiere (avance)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "active",
            "required": false,
            "in": "query",
            "description": "Filtre actif (ex. couleurs). true/false.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Champ de tri whitelisté côté service (ex. createdAt, name, orderDate).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Sens de tri",
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "includeVariants",
            "required": false,
            "in": "query",
            "description": "Inclure les variantes dans la liste (false par defaut)",
            "schema": {
              "default": false,
              "type": "boolean"
            }
          },
          {
            "name": "fields",
            "required": false,
            "in": "query",
            "description": "Niveau de detail des lignes retournees. compact = payload minimum, full = payload complet.",
            "schema": {
              "default": "full",
              "type": "string",
              "enum": [
                "compact",
                "full"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste paginee des produits",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponseDto"
                    },
                    {
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/ProductResponseDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lister les produits",
        "tags": [
          "Produits"
        ]
      },
      "post": {
        "operationId": "ProductsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Produit créé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Créer un produit",
        "tags": [
          "Produits"
        ]
      }
    },
    "/products/{id}": {
      "get": {
        "operationId": "ProductsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Produit trouvé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Récupérer un produit",
        "tags": [
          "Produits"
        ]
      },
      "patch": {
        "operationId": "ProductsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Produit mis à jour",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mettre à jour un produit",
        "tags": [
          "Produits"
        ]
      },
      "delete": {
        "operationId": "ProductsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Produit supprimé"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Supprimer un produit",
        "tags": [
          "Produits"
        ]
      }
    },
    "/products/{id}/technical-sheet": {
      "get": {
        "operationId": "ProductsController_getTechnicalSheet",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fiche technique du produit"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Récupérer la fiche technique atelier du produit",
        "tags": [
          "Produits"
        ]
      },
      "post": {
        "operationId": "ProductsController_upsertTechnicalSheet",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertProductTechnicalSheetDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Fiche technique enregistrée"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Créer ou mettre à jour la fiche technique (recette détaillée) du produit",
        "tags": [
          "Produits"
        ]
      }
    },
    "/products/{id}/images": {
      "get": {
        "operationId": "ProductsController_listImages",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Images du produit",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductImageResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lister les images d'un produit",
        "tags": [
          "Produits"
        ]
      },
      "post": {
        "operationId": "ProductsController_uploadImages",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "tag": {
                    "type": "string"
                  },
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                },
                "required": [
                  "files"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Images uploadées",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductImageResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Uploader plusieurs images produit (compression active)",
        "tags": [
          "Produits"
        ]
      }
    },
    "/products/{id}/images/{imageId}": {
      "get": {
        "operationId": "ProductsController_getImage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "imageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lire une image produit (décompression à la volée)",
        "tags": [
          "Produits"
        ]
      },
      "delete": {
        "operationId": "ProductsController_deleteImage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "imageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Image supprimée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteProductImageResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Supprimer une image produit",
        "tags": [
          "Produits"
        ]
      }
    },
    "/products/{id}/images/{imageId}/replace": {
      "post": {
        "operationId": "ProductsController_replaceImage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "imageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Image remplacée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplaceProductImageResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remplacer une image produit (nouvelle version)",
        "tags": [
          "Produits"
        ]
      }
    }
  },
  "info": {
    "title": "Raffia Weave Flow API",
    "description": "Fragment OpenAPI des modules contractualisés (search/reference-lookup/audit/templates/images). Document complet: GET /{SWAGGER_PATH}-json sur l’API vivante.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PaginatedResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "total": {
            "type": "number",
            "example": 125
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "pageSize": {
            "type": "number",
            "example": 20
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "pageSize"
        ]
      },
...
      "ProductResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "prd123"
          },
          "ref": {
            "type": "string",
            "example": "RAF-CAB-001"
          },
          "name": {
            "type": "string",
            "example": "Cabas Madagascar"
          },
          "basePrice": {
            "type": "number",
            "example": 89
          },
          "status": {
            "type": "string",
            "example": "ACTIVE"
          },
          "ownership": {
            "type": "string",
            "example": "COMPANY",
            "enum": [
              "COMPANY",
              "CLIENT"
            ],
            "description": "COMPANY = catalogue / vente. CLIENT = modèle propriété d'un client."
          },
          "ownerClientId": {
            "type": "object",
            "example": null,
            "nullable": true,
            "description": "Client propriétaire si ownership = CLIENT."
          }
        },
        "required": [
          "id",
          "ref",
          "name",
          "basePrice",
          "status",
          "ownership",
          "ownerClientId"
        ]
      },
...
      "UpsertProductTechnicalSheetDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Fiche Technique Cabas Madagascar - V1"
          },
          "instructions": {
            "type": "string",
            "example": "1) Préparer le raphia. 2) Crochet base. 3) Monter les poignées cuir. 4) Contrôle qualité."
          },
          "workshopNotes": {
            "type": "string",
            "example": "Respecter la tension de crochet sur les 3 premières rangées pour garder la forme du sac."
          },
          "elements": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UpsertProductTechnicalSheetElementDto"
            }
          }
        },
        "required": [
          "elements"
        ]
      },
      "ProductImageResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "img123"
          },
          "productId": {
            "type": "string",
            "example": "prd123"
          },
          "originalName": {
            "type": "string",
            "example": "cabas-front.jpg"
          },
          "mimeType": {
            "type": "string",
            "example": "image/jpeg"
          },
          "bucket": {
            "type": "object",
            "description": "Bucket MinIO quand le stockage objet est actif",
            "nullable": true,
            "example": "raffia-ged-raw"
          },
          "objectKey": {
            "type": "object",
            "description": "Clé objet MinIO",
            "nullable": true,
            "example": "admin/product/prd123/image/v1/cabas-front.jpg"
          },
          "storagePath": {
            "type": "object",
            "description": "Chemin disque local (fallback hors MinIO)",
            "nullable": true
          },
          "originalSize": {
            "type": "number",
            "description": "Taille originale avant compression (octets)",
            "example": 120000
          },
          "compressedSize": {
            "type": "number",
            "description": "Taille stockée après compression éventuelle (octets)",
            "example": 80000
          },
          "compressionAlgo": {
            "type": "string",
            "enum": [
              "NONE",
              "GZIP"
            ],
            "example": "GZIP"
          },
          "uploadedById": {
            "type": "object",
            "description": "Utilisateur ayant uploadé l’image",
            "nullable": true,
            "example": "usr123"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-07-01T10:00:00.000Z"
          },
          "version": {
            "type": "number",
            "description": "Version déduite de objectKey / storagePath",
            "example": 1
          },
          "decompressedUrl": {
            "type": "string",
            "description": "URL relative de lecture décompressée (présent sur findOne produit)",
            "example": "/products/prd123/images/img123"
          }
        },
        "required": [
          "id",
          "productId",
          "originalName",
          "mimeType",
          "originalSize",
          "compressedSize",
          "compressionAlgo",
          "createdAt"
        ]
      },
      "ReplaceProductImageResponseDto": {
        "type": "object",
        "properties": {
          "replacedImageId": {
            "type": "string",
            "example": "img123"
          },
          "newImageId": {
            "type": "string",
            "example": "img456"
          },
          "version": {
            "type": "number",
            "example": 2
          },
          "image": {
            "$ref": "#/components/schemas/ProductImageResponseDto"
          }
        },
        "required": [
          "replacedImageId",
          "newImageId",
          "version",
          "image"
        ]
      },
      "DeleteProductImageResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "img123"
          },
          "deleted": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "id",
          "deleted"
        ]
      },
...
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "ref": {
            "type": "string",
            "example": "S/123123",
            "description": "Reference produit. Si absente, elle est generee automatiquement depuis la categorie."
          },
          "name": {
            "type": "string",
            "example": "Cabas Madagascar"
          },
          "categoryId": {
            "type": "string",
            "example": "clx-category-id"
          },
          "basePrice": {
            "type": "number",
            "example": 89
          },
          "stockOnHand": {
            "type": "number",
            "example": 28
          },
          "status": {
            "type": "string",
            "example": "ACTIVE",
            "enum": [
              "ACTIVE",
              "LOW_STOCK",
              "OUT_OF_STOCK",
              "ARCHIVED"
            ]
          },
          "description": {
            "type": "string",
            "example": "Produit iconique atelier"
          },
          "ownership": {
            "type": "string",
            "example": "COMPANY",
            "enum": [
              "COMPANY",
              "CLIENT"
            ],
            "default": "COMPANY",
            "description": "COMPANY = produit entreprise (catalogue). CLIENT = modèle privé d'un client (hors catalogue)."
          },
          "ownerClientId": {
            "type": "string",
            "example": "clx-client-id",
            "description": "Obligatoire si ownership = CLIENT. Doit être null/omis si COMPANY."
          },
          "variants": {
            "description": "Variantes du produit (taille PM/MM/GM, coloris et dimensions par taille).",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateProductVariantDto"
            }
          }
        },
        "required": [
          "name",
          "categoryId",
          "basePrice"
        ]
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "ref": {
            "type": "string",
            "example": "S/123123",
            "description": "Reference produit. Si absente, elle est generee automatiquement depuis la categorie."
          },
          "name": {
            "type": "string",
            "example": "Cabas Madagascar"
          },
          "categoryId": {
            "type": "string",
            "example": "clx-category-id"
          },
          "basePrice": {
            "type": "number",
            "example": 89
          },
          "stockOnHand": {
            "type": "number",
            "example": 28
          },
          "status": {
            "type": "string",
            "example": "ACTIVE",
            "enum": [
              "ACTIVE",
              "LOW_STOCK",
              "OUT_OF_STOCK",
              "ARCHIVED"
            ]
          },
          "description": {
            "type": "string",
            "example": "Produit iconique atelier"
          },
          "ownership": {
            "type": "string",
            "example": "COMPANY",
            "enum": [
              "COMPANY",
              "CLIENT"
            ],
            "default": "COMPANY",
            "description": "COMPANY = produit entreprise (catalogue). CLIENT = modèle privé d'un client (hors catalogue)."
          },
          "ownerClientId": {
            "type": "string",
            "example": "clx-client-id",
            "description": "Obligatoire si ownership = CLIENT. Doit être null/omis si COMPANY."
          },
          "variants": {
            "description": "Variantes du produit (taille PM/MM/GM, coloris et dimensions par taille).",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateProductVariantDto"
            }
          }
        }
      }
    }
  }
}